+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ### [Unreleased][unreleased]
 
+### [1.4.0] - 2025-03-08
+#### Added
+- Added rawCommand
+
+### [1.3.1] - 2025-03-04
+#### Fixed
+- Fixed getConnection() for MultiConnectionPool
+
+### [1.3.0] - 2025-02-26
+####  Added
+- a few new commands: zincrby, hexpire, hexists, hstrlen
+- Multi connection support
+
+####  Fixed
+- several commands fixed to work with redis 6.*, 7.* and dragonfly
+
 ### [1.2.0] - 2024-04-12
 ####  Added
 - PHP 8.3 support
@@ -89,7 +105,10 @@
 #### Added
 - Wrapper for \Redis
 
-[unreleased]: https://github.com/lulco/redis-proxy/compare/1.2.0...HEAD
+[unreleased]: https://github.com/lulco/redis-proxy/compare/1.4.0...HEAD
+[1.4.0]: https://github.com/lulco/redis-proxy/compare/1.3.1...1.4.0
+[1.3.1]: https://github.com/lulco/redis-proxy/compare/1.3.0...1.3.1
+[1.3.0]: https://github.com/lulco/redis-proxy/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/lulco/redis-proxy/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/lulco/redis-proxy/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/lulco/redis-proxy/compare/0.7.1...1.0.0